- Generalized MemoryObject.getType to not necessarily be an integer (we still use integers in practice) - Removed automatic initialization for static MemoryObjects. Instead, the ProgramBuilder will do so explicitly when building from litmus code.

+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/EventFactory.java
@@ -131,7 +131,7 @@ public class EventFactory {
         //TODO: We simplify here because virtual aliasing currently fails when pointer arithmetic is involved
         // meaning that <addr> and <addr + 0> are treated differently.
         final Expression address = offset == 0 ? base :
-                expressions.makeAdd(base, expressions.makeValue(offset, base.getType()));
+                expressions.makeAdd(base, expressions.makeValue(offset, (IntegerType) base.getType()));
         return new Init(base, offset, address);
     }
